third_party/blink/web_tests/css3/flexbox/flex-align-baseline-expected.txt

PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetLeft is lastChild.offsetLeft
PASS firstChild.offsetTop is lastChild.offsetTop
PASS firstChild.offsetTop is lastChild.offsetTop
PASS successfullyParsed is true

TEST COMPLETE

horizontal-tb ltr row
horizontal-tb rtl row
horizontal-tb ltr column
horizontal-tb rtl column
horizontal-tb ltr row-reverse
horizontal-tb rtl row-reverse
horizontal-tb ltr column-reverse
horizontal-tb rtl column-reverse
vertical-lr ltr row
vertical-lr rtl row
vertical-lr ltr column
vertical-lr rtl column
vertical-lr ltr row-reverse
vertical-lr rtl row-reverse
vertical-lr ltr column-reverse
vertical-lr rtl column-reverse
vertical-rl ltr row
vertical-rl rtl row
vertical-rl ltr column
vertical-rl rtl column
vertical-rl ltr row-reverse
vertical-rl rtl row-reverse
vertical-rl ltr column-reverse
vertical-rl rtl column-reverse
